feat: rename torture-generator to go2nix-testgen, use DAG mode
Rename packages/torture-generator to packages/go2nix-testgen and
switch from buildGoModule to buildGoApplicationDAGMode (dogfooding).
Module path: github.com/numtide/go2nix/testgen.
