+++ packages/go2nix-testgen/go.mod
@@ -1,0 +1,3 @@
+module github.com/numtide/go2nix/testgen
+
+go 1.25
