+++ packages/go2nix-testgen/go2nix.toml
@@ -1,0 +1,3 @@
+# go2nix lockfile v2. Generated by go2nix. Do not edit.
+
+[mod]
