--- packages/torture-generator/default.nix
@@ -1,1 +1,0 @@
-{ pkgs, ... }: pkgs.callPackage ./package.nix { }
