--- packages/torture-generator/go.mod
@@ -1,3 +1,0 @@
-module github.com/numtide/go2nix/torture-generator
-
-go 1.25
